A xaringan template for presentations

An opinionated template

utilitr_presentation(
  css = NULL,
  nature = list(highlightStyle = "github", highlightLines = TRUE,
    countIncrementalSlides = TRUE),
  self_contained = TRUE,
  lib_dir = "libs",
  ...
)

Arguments

css

A vector of CSS file paths. Two default CSS files (default.css and default-fonts.css) are provided in this package, which was borrowed from https://remarkjs.com. If the character vector css contains a value that does not end with .css, it is supposed to be a built-in CSS file in this package, e.g., for css = c('default', 'extra.css')), it means default.css in this package and a user-provided extra.css. To find out all built-in CSS files, use xaringan:::list_css().

nature

(Nature transformation) A list of configurations to be passed to remark.create(), e.g. list(ratio = '16:9', navigation = list(click = TRUE)); see https://github.com/gnab/remark/wiki/Configuration. Besides the options provided by remark.js, you can also set autoplay to a number (the number of milliseconds) so the slides will be played every autoplay milliseconds; alternatively, autoplay can be a list of the form list(interval = N, loop = TRUE), so the slides will go to the next page every N milliseconds, and optionally go back to the first page to restart the play when loop = TRUE. You can also set countdown to a number (the number of milliseconds) to include a countdown timer on each slide. If using autoplay, you can optionally set countdown to TRUE to include a countdown equal to autoplay. To alter the set of classes applied to the title slide, you can optionally set titleSlideClass to a vector of classes; the default is c("center", "middle", "inverse").

self_contained

Whether to produce a self-contained HTML file by embedding all external resources into the HTML file. See the ‘Note’ section below.

lib_dir

JS dependencies librairies location

...

For tsukuyomi(), arguments passed to moon_reader(); for moon_reader(), arguments passed to rmarkdown::html_document().